to merger master

+++ packages/context/src/lib/context-manager/shared/context.service.ts
@@ -2,9 +2,18 @@ import { Injectable, Optional } from '@angular/core';
 import { HttpClient, HttpErrorResponse } from '@angular/common/http';
 
 import { BehaviorSubject, Observable, of } from 'rxjs';
-import { map, tap, catchError, debounceTime, flatMap } from 'rxjs/operators';
+import {
+  map,
+  tap,
+  catchError,
+  debounceTime,
+  mergeMap,
+  first
+} from 'rxjs/operators';
 
 import olPoint from 'ol/geom/Point';
+import GeoJSON from 'ol/format/GeoJSON';
+import Cluster from 'ol/source/Cluster';
 
 import { Tool } from '@igo2/common';
 import { uuid, ObjectUtils } from '@igo2/utils';
@@ -18,7 +27,7 @@ import {
 } from '@igo2/core';
 
 import { AuthService } from '@igo2/auth';
-import { IgoMap } from '@igo2/geo';
+import type { IgoMap, Layer } from '@igo2/geo';
 
 import { TypePermission } from './context.enum';
 import {
@@ -27,7 +36,8 @@ import {
   Context,
   DetailedContext,
   ContextMapView,
-  ContextPermission
+  ContextPermission,
+  ContextProfils
 } from './context.interface';
 
 @Injectable({
@@ -38,6 +48,7 @@ export class ContextService {
   public contexts$ = new BehaviorSubject<ContextsList>({ ours: [] });
   public defaultContextId$ = new BehaviorSubject<string>(undefined);
   public editedContext$ = new BehaviorSubject<DetailedContext>(undefined);
+  public importedContext: Array<DetailedContext> = [];
   private mapViewFromRoute: ContextMapView = {};
   private options: ContextServiceOptions;
   private baseUrl: string;
@@ -46,6 +57,15 @@ export class ContextService {
   // Until the ContextService is completely refactored, this is needed
   // to track the current tools
   private tools: Tool[];
+  private toolbar: string[];
+
+  get defaultContextUri(): string {
+    return this._defaultContextUri || this.options.defaultContextUri;
+  }
+  set defaultContextUri(uri: string) {
+    this._defaultContextUri = uri;
+  }
+  private _defaultContextUri: string;
 
   constructor(
     private http: HttpClient,
@@ -68,19 +88,28 @@ export class ContextService {
 
     this.readParamsFromRoute();
 
-    this.authService.authenticate$.subscribe(authenticated => {
-      const contexts$$ = this.contexts$.subscribe(contexts => {
-        if (contexts$$) {
-          contexts$$.unsubscribe();
+    this.authService.authenticate$.subscribe((authenticated) => {
+      if (authenticated && this.baseUrl) {
+        this.get().subscribe((contexts) => {
           this.handleContextsChange(contexts);
-        }
-      });
-      this.loadContexts();
+        });
+      } else {
+        this.contexts$.pipe(first()).subscribe((contexts) => {
+          this.handleContextsChange(contexts);
+        });
+        this.loadContexts();
+      }
     });
   }
 
-  get(): Observable<ContextsList> {
-    const url = this.baseUrl + '/contexts';
+  get(permissions?: string[], hidden?: boolean): Observable<ContextsList> {
+    let url = this.baseUrl + '/contexts';
+    if (permissions && this.authService.authenticated) {
+      url += '?permission=' + permissions.join();
+      if (hidden) {
+        url += '&hidden=true';
+      }
+    }
     return this.http.get<ContextsList>(url);
   }
 
@@ -92,7 +121,7 @@ export class ContextService {
   getDetails(id: string): Observable<DetailedContext> {
     const url = `${this.baseUrl}/contexts/${id}/details`;
     return this.http.get<DetailedContext>(url).pipe(
-      catchError(res => {
+      catchError((res) => {
         return this.handleError(res, id);
       })
     );
@@ -101,10 +130,18 @@ export class ContextService {
   getDefault(): Observable<DetailedContext> {
     const url = this.baseUrl + '/contexts/default';
     return this.http.get<DetailedContext>(url).pipe(
-      tap(context => {
+      tap((context) => {
         this.defaultContextId$.next(context.id);
       })
     );
+  }
+
+  getProfilByUser(): Observable<ContextProfils[]> {
+    if (this.baseUrl) {
+      const url = this.baseUrl + '/profils?';
+      return this.http.get<ContextProfils[]>(url);
+    }
+    return of([]);
   }
 
   setDefault(id: string): Observable<any> {
@@ -112,15 +149,31 @@ export class ContextService {
     return this.http.post(url, { defaultContextId: id });
   }
 
-  delete(id: string): Observable<void> {
+  hideContext(id: string) {
+    const url = this.baseUrl + '/contexts/' + id + '/hide';
+    return this.http.post(url, {});
+  }
+
+  showContext(id: string) {
+    const url = this.baseUrl + '/contexts/' + id + '/show';
+    return this.http.post(url, {});
+  }
+
+  delete(id: string, imported = false): Observable<void> {
+    const contexts: ContextsList = { ours: [] };
+    Object.keys(this.contexts$.value).forEach(
+      (key) =>
+        (contexts[key] = this.contexts$.value[key].filter((c) => c.id !== id))
+    );
+
+    if (imported) {
+      this.importedContext = this.importedContext.filter((c) => c.id !== id);
+      return of(this.contexts$.next(contexts));
+    }
+
     const url = this.baseUrl + '/contexts/' + id;
     return this.http.delete<void>(url).pipe(
-      tap(res => {
-        const contexts: ContextsList = { ours: [] };
-        Object.keys(this.contexts$.value).forEach(
-          key =>
-            (contexts[key] = this.contexts$.value[key].filter(c => c.id !== id))
-        );
+      tap((res) => {
         this.contexts$.next(contexts);
       })
     );
@@ -129,13 +182,13 @@ export class ContextService {
   create(context: DetailedContext): Observable<Context> {
     const url = this.baseUrl + '/contexts';
     return this.http.post<Context>(url, JSON.stringify(context)).pipe(
-      map(contextCreated => {
+      map((contextCreated) => {
         if (this.authService.authenticated) {
           contextCreated.permission = TypePermission[TypePermission.write];
         } else {
           contextCreated.permission = TypePermission[TypePermission.read];
         }
-        this.contexts$.value.ours.push(contextCreated);
+        this.contexts$.value.ours.unshift(contextCreated);
         this.contexts$.next(this.contexts$.value);
         return contextCreated;
       })
@@ -145,9 +198,9 @@ export class ContextService {
   clone(id: string, properties = {}): Observable<Context> {
     const url = this.baseUrl + '/contexts/' + id + '/clone';
     return this.http.post<Context>(url, JSON.stringify(properties)).pipe(
-      map(contextCloned => {
+      map((contextCloned) => {
         contextCloned.permission = TypePermission[TypePermission.write];
-        this.contexts$.value.ours.push(contextCloned);
+        this.contexts$.value.ours.unshift(contextCloned);
         this.contexts$.next(this.contexts$.value);
         return contextCloned;
       })
@@ -193,7 +246,7 @@ export class ContextService {
     return this.http
       .post<ContextPermission[]>(url, JSON.stringify(association))
       .pipe(
-        catchError(res => {
+        catchError((res) => {
           return [this.handleError(res, undefined, true)];
         })
       );
@@ -221,7 +274,7 @@ export class ContextService {
   getLocalContext(uri: string): Observable<DetailedContext> {
     const url = this.getPath(`${uri}.json`);
     return this.http.get<DetailedContext>(url).pipe(
-      flatMap(res => {
+      mergeMap((res) => {
         if (!res.base) {
           return of(res);
         }
@@ -235,37 +288,39 @@ export class ContextService {
               .reverse()
               .filter(
                 (l, index, self) =>
-                  !l.id || self.findIndex(l2 => l2.id === l.id) === index
+                  !l.id || self.findIndex((l2) => l2.id === l.id) === index
               )
               .reverse();
             resMerge.toolbar = res.toolbar || resBase.toolbar;
+            resMerge.message = res.message || resBase.message;
             resMerge.tools = (res.tools || [])
               .concat(resBase.tools || [])
               .filter(
                 (t, index, self) =>
-                  self.findIndex(t2 => t2.name === t.name) === index
+                  self.findIndex((t2) => t2.name === t.name) === index
               );
             return resMerge;
           }),
-          catchError(err => {
+          catchError((err) => {
             return this.handleError(err, uri);
           })
         );
       }),
-      catchError(err2 => {
+      catchError((err2) => {
         return this.handleError(err2, uri);
       })
     );
   }
 
-  loadContexts() {
+  loadContexts(permissions?: string[], hidden?: boolean) {
     let request;
     if (this.baseUrl) {
-      request = this.get();
+      request = this.get(permissions, hidden);
     } else {
       request = this.getLocalContexts();
     }
-    request.subscribe(contexts => {
+    request.subscribe((contexts) => {
+      contexts.ours = this.importedContext.concat(contexts.ours);
       this.contexts$.next(contexts);
     });
   }
@@ -275,26 +330,26 @@ export class ContextService {
       if (!direct && this.baseUrl && this.authService.authenticated) {
         this.getDefault().subscribe(
           (_context: DetailedContext) => {
-            this.options.defaultContextUri = _context.uri;
+            this.defaultContextUri = _context.uri;
             this.addContextToList(_context);
             this.setContext(_context);
           },
           () => {
             this.defaultContextId$.next(undefined);
-            this.loadContext(this.options.defaultContextUri);
+            this.loadContext(this.defaultContextUri);
           }
         );
       } else {
-        this.loadContext(this.options.defaultContextUri);
+        this.loadContext(this.defaultContextUri);
       }
     };
 
     if (this.route && this.route.options.contextKey) {
-      this.route.queryParams.pipe(debounceTime(100)).subscribe(params => {
+      this.route.queryParams.pipe(debounceTime(100)).subscribe((params) => {
         const contextParam = params[this.route.options.contextKey as string];
         let direct = false;
         if (contextParam) {
-          this.options.defaultContextUri = contextParam;
+          this.defaultContextUri = contextParam;
           direct = true;
         }
         loadFct(direct);
@@ -306,20 +361,24 @@ export class ContextService {
 
   loadContext(uri: string) {
     const context = this.context$.value;
+
     if (context && context.uri === uri) {
       return;
     }
 
-    const contexts$$ = this.getContextByUri(uri).subscribe(
-      (_context: DetailedContext) => {
-        contexts$$.unsubscribe();
-        this.addContextToList(_context);
-        this.setContext(_context);
-      },
-      err => {
-        contexts$$.unsubscribe();
-      }
-    );
+    this.getContextByUri(uri)
+      .pipe(first())
+      .subscribe(
+        (_context: DetailedContext) => {
+          this.addContextToList(_context);
+          this.setContext(_context);
+        },
+        (err) => {
+          if (uri !== this.options.defaultContextUri) {
+            this.loadContext(this.options.defaultContextUri);
+          }
+        }
+      );
   }
 
   setContext(context: DetailedContext) {
@@ -352,7 +411,7 @@ export class ContextService {
     this.editedContext$.next(context);
   }
 
-  getContextFromMap(igoMap: IgoMap): DetailedContext {
+  getContextFromMap(igoMap: IgoMap, empty?: boolean): DetailedContext {
     const view = igoMap.ol.getView();
     const proj = view.getProjection().getCode();
     const center: any = new olPoint(view.getCenter()).transform(
@@ -368,16 +427,27 @@ export class ContextService {
         view: {
           center: center.getCoordinates(),
           zoom: view.getZoom(),
-          projection: proj
+          projection: proj,
+          maxZoomOnExtent: igoMap.viewController.maxZoomOnExtent
         }
       },
       layers: [],
       tools: []
     };
 
-    const layers = igoMap.layers$
-      .getValue()
-      .sort((a, b) => a.zIndex - b.zIndex);
+    let layers = [];
+    if (empty === true) {
+      layers = igoMap.layers$
+        .getValue()
+        .filter(
+          (lay) =>
+            lay.baseLayer === true ||
+            lay.options.id === 'searchPointerSummaryId'
+        )
+        .sort((a, b) => a.zIndex - b.zIndex);
+    } else {
+      layers = igoMap.layers$.getValue().sort((a, b) => a.zIndex - b.zIndex);
+    }
 
     let i = 0;
     for (const l of layers) {
@@ -401,15 +471,127 @@ export class ContextService {
       }
     }
 
-    context.tools = this.tools.map(tool => {
+    context.tools = this.tools.map((tool) => {
       return { id: String(tool.id), global: tool.global };
     });
 
     return context;
   }
 
+  getContextFromLayers(
+    igoMap: IgoMap,
+    layers: Layer[],
+    name: string
+  ): DetailedContext {
+    const currentContext = this.context$.getValue();
+    const view = igoMap.ol.getView();
+    const proj = view.getProjection().getCode();
+    const center: any = new olPoint(view.getCenter()).transform(
+      proj,
+      'EPSG:4326'
+    );
+
+    const context = {
+      uri: name,
+      title: name,
+      map: {
+        view: {
+          center: center.getCoordinates(),
+          zoom: view.getZoom(),
+          projection: proj
+        }
+      },
+      layers: [],
+      toolbar: [],
+      tools: [],
+      extraFeatures: []
+    };
+
+    const currentLayers = igoMap.layers$.getValue();
+    context.layers = currentLayers
+      .filter((l) => l.baseLayer)
+      .map((l) => {
+        return {
+          baseLayer: true,
+          sourceOptions: l.options.sourceOptions,
+          title: l.options.title,
+          visible: l.visible
+        };
+      });
+
+    layers.forEach((layer) => {
+      const layerFound = currentContext.layers.find(
+        (contextLayer) =>
+          layer.id === contextLayer.source.id && !contextLayer.baseLayer
+      );
+
+      if (layerFound) {
+        let layerStyle = layerFound[`style`];
+        if (layerFound[`styleByAttribute`]) {
+          layerStyle = undefined;
+        } else if (layerFound[`clusterBaseStyle`]) {
+          layerStyle = undefined;
+          delete layerFound.sourceOptions[`source`];
+          delete layerFound.sourceOptions[`format`];
+        }
+        const opts = {
+          baseLayer: layerFound.baseLayer,
+          title: layer.options.title,
+          zIndex: layer.zIndex,
+          styleByAttribute: layerFound[`styleByAttribute`],
+          clusterBaseStyle: layerFound[`clusterBaseStyle`],
+          style: layerStyle,
+          clusterParam: layerFound[`clusterParam`],
+          visible: layer.visible,
+          opacity: layer.opacity,
+          sourceOptions: layerFound.sourceOptions
+        };
+        context.layers.push(opts);
+      } else {
+        if (layer.ol.type !== 'VECTOR') {
+          const catalogLayer = layer.options;
+          catalogLayer.zIndex = layer.zIndex;
+          delete catalogLayer.source;
+          context.layers.push(catalogLayer);
+        } else {
+          let features;
+          const writer = new GeoJSON();
+          if (layer.ol.getSource() instanceof Cluster) {
+            features = writer.writeFeatures(
+              layer.ol.getSource().getSource().getFeatures(),
+              {
+                dataProjection: 'EPSG:4326',
+                featureProjection: 'EPSG:3857'
+              }
+            );
+          } else {
+            features = writer.writeFeatures(
+              layer.ol.getSource().getFeatures(),
+              {
+                dataProjection: 'EPSG:4326',
+                featureProjection: 'EPSG:3857'
+              }
+            );
+          }
+          features = JSON.parse(features);
+          features.name = layer.options.title;
+          context.extraFeatures.push(features);
+        }
+      }
+    });
+
+    context.toolbar = this.toolbar;
+    context.tools = this.tools;
+
+    return context;
+  }
+
   setTools(tools: Tool[]) {
     this.tools = tools;
+  }
+
+  setToolbar(toolbar: string[]) {
+    this.toolbar = toolbar;
   }
 
   private handleContextMessage(context: DetailedContext) {
@@ -432,7 +614,7 @@ export class ContextService {
     if (this.baseUrl) {
       let contextToLoad;
       for (const key of Object.keys(this.contexts$.value)) {
-        contextToLoad = this.contexts$.value[key].find(c => {
+        contextToLoad = this.contexts$.value[key].find((c) => {
           return c.uri === uri;
         });
         if (contextToLoad) {
@@ -440,12 +622,35 @@ export class ContextService {
         }
       }
 
+      if (contextToLoad && contextToLoad.imported) {
+        return of(contextToLoad);
+      }
+
       // TODO : use always id or uri
       const id = contextToLoad ? contextToLoad.id : uri;
       return this.getDetails(id);
     }
 
-    return this.getLocalContext(uri);
+    const importedContext = this.contexts$.value.ours.find((currentContext) => {
+      return currentContext.uri === uri && currentContext.imported === true;
+    });
+
+    if (importedContext) {
+      return of(importedContext);
+    } else {
+      return this.getLocalContext(uri);
+    }
+  }
+
+  getContextLayers(igoMap: IgoMap) {
+    const layers: Layer[] = [];
+    const mapLayers = igoMap.layers$.getValue();
+    mapLayers.forEach((layer) => {
+      if (!layer.baseLayer && layer.options.id !== 'searchPointerSummaryId') {
+        layers.push(layer);
+      }
+    });
+    return layers;
   }
 
   private readParamsFromRoute() {
@@ -453,7 +658,7 @@ export class ContextService {
       return;
     }
 
-    this.route.queryParams.subscribe(params => {
+    this.route.queryParams.subscribe((params) => {
       const centerKey = this.route.options.centerKey;
       if (centerKey && params[centerKey as string]) {
         const centerParams = params[centerKey as string];
@@ -486,7 +691,7 @@ export class ContextService {
     uri: string,
     permissionError?: boolean
   ): Message[] {
-    const context = this.contexts$.value.ours.find(obj => obj.uri === uri);
+    const context = this.contexts$.value.ours.find((obj) => obj.uri === uri);
     const titleContext = context ? context.title : uri;
     error.error.title = this.languageService.translate.instant(
       'igo.context.contextManager.invalid.title'
@@ -545,6 +750,7 @@ export class ContextService {
         scope: context.scope,
         permission: TypePermission[TypePermission.read]
       };
+
       if (this.contexts$.value && this.contexts$.value.public) {
         this.contexts$.value.public.push(contextSimplifie);
         this.contexts$.next(this.contexts$.value);
@@ -562,7 +768,7 @@ export class ContextService {
     for (const key of Object.keys(contexts)) {
       const value = contexts[key];
       found = value.find(
-        c =>
+        (c) =>
           (context.id && c.id === context.id) ||
           (context.uri && c.uri === context.uri)
       );
